backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2
pydantic-settings>=2
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
bcrypt>=4.1
PyJWT>=2.8
pyotp>=2.9
cryptography>=42.0.0
